Lot more tests! - Tests for converting legacy terrain types - Tests for initialization to sane values
Fixes:
- cCreature no longer initializes spec1 and spec2 to 0. This probably didn't cause bugs, but who knows...
- In fact, cCreature no longer explicitly initializes anything in its cTownperson superclass. That's what the superclass constructor is for after all.
- Relatedly, cTownperson now defaults to a facial graphic of -1. Also a docile attitude.
- iLiving defaults to ap 0, direction here.
- Fix Change When Step terrains to properly support not having a sound while still allowing an arbitrary number of custom sounds.
- Add support for a lack of sound to Change When Used terrains, including at conversion time (original game supported it but OBoE was forcing the sound to a door sound).
- Fix conversion of crumbling terrain types (old conversion was allowing quickfire to destroy them)

Other:
- New addAttack method in cMonster.
- Change snd_num_t to signed, because it's been getting annoying that I can't easily use -1 to mean "no sound".

+++ src/tools/soundtool.hpp
@@ -11,7 +11,7 @@
 
 #include <SFML/Audio.hpp>
 
-typedef unsigned short snd_num_t;
+typedef signed int snd_num_t;
 void init_snd_tool();
 bool sound_going(snd_num_t which_s);
 void play_sound(short which, short how_many_times = 1);
